Phantom Bonds and fragment library update
Main changes here are adding phantom bonds functionality.  This allows find fragment to be able to see how many bonds an atom is bonded to.  Used to distinguish between primary/secondary/etc... amines and esters and carboxylic acids for example.  If you want a fragment to do a phantom bonds check on an atom, make an extra bond from that atom to either W, X, Y, or Z (1, 2, 3, and 4 bonds respectively).  This will assign that value to the atom_info's phantom_bonds attribute in the atom_map.  When find_fragment is checking an atom and the atom_info's phantom bonds attribute is > 0, it performs a check to make sure that atom has the required number of bonds.  For a tertiary amine, the SMILES string is "NY", which will check to make sure that nitrogen atom is bonded to three other atoms.   "W" can be useful for signifying a terminus: "tert-butyl": "YC(CW)(CW)(CW)".  This ensures that each of the methyl groups on a tert-butyl group are not longer alkyl chains (whereas just "C(C)(C)(C)" would be found for any tertiary carbon.  Made some updates to the fragment library (added additional resonance structures for some fragments) and improved the fragment hierarchy