check_bond function and adding Q, R elements and & bond
added a check_bond function to see if the current bond is the same as the map bond.  Wrapping this in a function allows for defining a more complex bond checking system.  now fragments in the library can have Q (any heteroatom) and R (any element) atoms.  Bonds can be defined as &, which is interperted as any bond.  This will allow substructure searches with anonymous bonds and atoms (to generalize the search).