Phantom Atom update
Added phantom atom functionality.  This allows there to be portions of fragment strings that can be searched for, but aren't marked as discovered.  This is marked by putting the phantom part of the fragment string in curly brackets "{}".  For example, benzoic acid is represented by this smiles string: OC(C1=CC=CC=C1)=O.  If the string is rewritten as with the benzene portion in curly brackets:
OC({C1=CC=CC=C1})=O, then if a benzoic acid fragment is found, the benzene portion will not be marked as discovered.  So a benzene fragment could also be found in this portion of the molecule.  It doesn't matter where you put the curly brackets so long as all the element symbols you want to mark as phantom atoms are within the curly brackets.

The next update will allow phantom atoms to be found in portions of the molecule that have already been discovered.

Also fixed a bug with currently_visited that occasinally removed atoms from the dictionary ->
Only need to pop atoms off currently_visited after failed branch or failed midway fork.