Phantom atoms can be found in discovered portions of molecule
Updated check_bond:
now check_bond does a bond.atom.discovered check.  It will return False unless the phantom_atom attribute for the bond.atom is True.  i.e. if the atom_info for the atom in the bond being checked is deonated as a phantom atom, it can be found if atom of the actual molecule has already been discovered.

The atom_info for bond.atom (i.e. the atom that the current one is bonded to) is now passed as an argument to check_bond.  For check_bond_atoms or midway_fork, it is passed as the next atom_info in the branch's sequence: (branch_sequence[index + 1]).  For a branch point, where check_bond is used to assess whether a certain path is viable for each branch, the atom_info for the first atom in the branch's sequence is passed as the arg (i.e. if the first atom_info in the branch sequence is a phantom atom, the branch can be tried on a path where that atom has already been discovered).